Project_Selector/requirements.txt
GitPython
wxPython
